+++ snowchange.py
@@ -1,18 +1,34 @@
 import os
+import string
 import re
 import argparse
+import json
 import time
 import hashlib
 import snowflake.connector
 
 # Set a few global variables here
-_snowchange_version = '2.1.0'
+_snowchange_version = '2.2.0'
 _metadata_database_name = 'METADATA'
 _metadata_schema_name = 'SNOWCHANGE'
 _metadata_table_name = 'CHANGE_HISTORY'
 
+# Define the Jinja expression template class
+# snowchange uses Jinja style variable references of the form "{{ variablename }}"
+# See https://jinja.palletsprojects.com/en/2.11.x/templates/
+# Variable names follow Python variable naming conventions
+class JinjaExpressionTemplate(string.Template):
+    delimiter = '{{ '
+    pattern = r'''
+    \{\{[ ](?:
+    (?P<escaped>\{\{)|
+    (?P<named>[_A-Za-z][_A-Za-z0-9]*)[ ]\}\}|
+    (?P<braced>[_A-Za-z][_A-Za-z0-9]*)[ ]\}\}|
+    (?P<invalid>)
+    )
+    '''
 
-def snowchange(root_folder, snowflake_account, snowflake_region, snowflake_user, snowflake_role, snowflake_warehouse, change_history_table_override, autocommit, verbose):
+def snowchange(root_folder, snowflake_account, snowflake_user, snowflake_role, snowflake_warehouse, change_history_table_override, vars, autocommit, verbose):
   if "SNOWSQL_PWD" not in os.environ:
     raise ValueError("The SNOWSQL_PWD environment variable has not been defined")
 
@@ -22,13 +38,14 @@ def snowchange(root_folder, snowflake_account, snowflake_region, snowflake_user,
 
   print("snowchange version: %s" % _snowchange_version)
   print("Using root folder %s" % root_folder)
+  print("Using variables %s" % vars)
+  print("Using Snowflake account %s" % snowflake_account)
 
   # TODO: Is there a better way to do this without setting environment variables?
   os.environ["SNOWFLAKE_ACCOUNT"] = snowflake_account
   os.environ["SNOWFLAKE_USER"] = snowflake_user
   os.environ["SNOWFLAKE_ROLE"] = snowflake_role
   os.environ["SNOWFLAKE_WAREHOUSE"] = snowflake_warehouse
-  os.environ["SNOWFLAKE_REGION"] = snowflake_region
   os.environ["SNOWFLAKE_AUTHENTICATOR"] = 'snowflake'
 
   scripts_skipped = 0
@@ -72,7 +89,7 @@ def snowchange(root_folder, snowflake_account, snowflake_region, snowflake_user,
       continue
 
     print("Applying change script %s" % script['script_name'])
-    apply_change_script(script, change_history_table, autocommit, verbose)
+    apply_change_script(script, vars, change_history_table, autocommit, verbose)
     scripts_applied += 1
 
   print("Successfully applied %d change scripts (skipping %d)" % (scripts_applied, scripts_skipped))
@@ -128,7 +145,6 @@ def execute_snowflake_query(snowflake_database, query, autocommit, verbose):
     role = os.environ["SNOWFLAKE_ROLE"],
     warehouse = os.environ["SNOWFLAKE_WAREHOUSE"],
     database = snowflake_database,
-    region = os.environ["SNOWFLAKE_REGION"],
     authenticator = os.environ["SNOWFLAKE_AUTHENTICATOR"],
     password = os.environ["SNOWSQL_PWD"]
   )
@@ -200,7 +216,7 @@ def fetch_change_history(change_history_table, autocommit, verbose):
 
   return change_history
 
-def apply_change_script(script, change_history_table, autocommit, verbose):
+def apply_change_script(script, vars, change_history_table, autocommit, verbose):
   # First read the contents of the script
   with open(script['script_full_path'],'r') as content_file:
     content = content_file.read().strip()
@@ -210,6 +226,9 @@ def apply_change_script(script, change_history_table, autocommit, verbose):
   checksum = hashlib.sha224(content.encode('utf-8')).hexdigest()
   execution_time = 0
   status = 'Success'
+
+  # Replace any variables used in the script content
+  content = replace_variables_references(content, vars, verbose)
 
   # Execute the contents of the script
   if len(content) > 0:
@@ -222,18 +241,25 @@ def apply_change_script(script, change_history_table, autocommit, verbose):
   query = "INSERT INTO {0}.{1} (VERSION, DESCRIPTION, SCRIPT, SCRIPT_TYPE, CHECKSUM, EXECUTION_TIME, STATUS, INSTALLED_BY, INSTALLED_ON) values ('{2}','{3}','{4}','{5}','{6}',{7},'{8}','{9}',CURRENT_TIMESTAMP);".format(change_history_table['schema_name'], change_history_table['table_name'], script['script_version'], script['script_description'], script['script_name'], script['script_type'], checksum, execution_time, status, os.environ["SNOWFLAKE_USER"])
   execute_snowflake_query(change_history_table['database_name'], query, autocommit, verbose)
 
+# This method will throw an error if there are any leftover variables in the change script
+# Since a leftover variable in the script isn't valid SQL, and will fail when run it's
+# better to throw an error here and have the user fix the problem ahead of time.
+def replace_variables_references(content, vars, verbose):
+  t = JinjaExpressionTemplate(content)
+  return t.substitute(vars)
+
 
 if __name__ == '__main__':
   parser = argparse.ArgumentParser(prog = 'python snowchange.py', description = 'Apply schema changes to a Snowflake account. Full readme at https://github.com/jamesweakley/snowchange', formatter_class = argparse.RawTextHelpFormatter)
   parser.add_argument('-f','--root-folder', type = str, default = ".", help = 'The root folder for the database change scripts')
-  parser.add_argument('-a', '--snowflake-account', type = str, help = 'The name of the snowflake account (e.g. ly12345)', required = True)
-  parser.add_argument('--snowflake-region', type = str, help = 'The name of the snowflake region (e.g. ap-southeast-2)', required = True)
+  parser.add_argument('-a', '--snowflake-account', type = str, help = 'The name of the snowflake account (e.g. abc123.east-us-2.azure)', required = True)
   parser.add_argument('-u', '--snowflake-user', type = str, help = 'The name of the snowflake user (e.g. DEPLOYER)', required = True)
   parser.add_argument('-r', '--snowflake-role', type = str, help = 'The name of the role to use (e.g. DEPLOYER_ROLE)', required = True)
   parser.add_argument('-w', '--snowflake-warehouse', type = str, help = 'The name of the warehouse to use (e.g. DEPLOYER_WAREHOUSE)', required = True)
   parser.add_argument('-c', '--change-history-table', type = str, help = 'Used to override the default name of the change history table (e.g. METADATA.SNOWCHANGE.CHANGE_HISTORY)', required = False)
+  parser.add_argument('--vars', type = json.loads, help = 'Define values for the variables to replaced in change scripts, given in JSON format (e.g. {"variable1": "value1", "variable2": "value2"})', required = False)
   parser.add_argument('-ac', '--autocommit', action='store_true')
   parser.add_argument('-v','--verbose', action='store_true')
   args = parser.parse_args()
 
-  snowchange(args.root_folder, args.snowflake_account, args.snowflake_region, args.snowflake_user, args.snowflake_role, args.snowflake_warehouse, args.change_history_table, args.autocommit, args.verbose)
+  snowchange(args.root_folder, args.snowflake_account, args.snowflake_user, args.snowflake_role, args.snowflake_warehouse, args.change_history_table, args.vars, args.autocommit, args.verbose)